Use Jinja2 for jobs and fix stats job
The template language used by jenkins job builder is frankly rubbish.
This passes the job definitions through jinja2 to allow us to all the
goodness jinja has to offer. The particular problem being fixed here is
that we need the environment to be uppercase in the environment variable
but lowercase elsewhere.