Add conditional slack notifications
Notify slack if the job fails or if the job succeeds and was triggered
by and upstream job. This means that we don't get a continuous stream of
notifications saying that the tests are fine but we do get explicit
confirmation that our releases haven't caused them to break.

This also adds the SSP functional tests in.