Use property file to get release version for slack/job name
Updated aws release script creates a release_properties.out file
with RELEASE_NAME=version pair format, which is supported by
name setter and build trigger plugins.

This allows us to always have a release number in both job names and
slack notifications.

This requires an additional post-build step to wipe the contents of
the properties file. Otherwise if the file is left over after the
previous job run the name setter plugin will pick it up when the job
starts - adding a previous release version to the current job name
until the job is finished (at which point it's renamed with the
correct relese version). Wiping the file contents sets the running
job name without release name and adds it once the job is finished.

Since script post-build tasks seem to run before the triggered jobs
despite the given order of the actions we copy the current property
file with a different name to use in notification steps.