Remove unneeded checkout & refine describe command
The tag sent to `push_to_github_pages.sh` is
checked out in that script so it's unnecessary to
do it here.

The output of `git describe` is usually just the
name of the tag for the HEAD commit of the current
branch. This is the behaviour in most cases.

If the tag points to a commit which is not the
HEAD however (which can happen if some commits
have been made since the tag was created) then the
output will contain other information about these
commits. For example:

<tag>-<number of commits since the tag>-<SHA of
last commit>

ie

v11.2.0-2-ds3j34kb322337ye3y2ei3323ekrjb

The display of the SHA is controlled by the
`--abbrev` flag and set to 0 it will remove the
extra information completely.

We use this method here to ensure only the tag
name is used.