Add update-tls-cert Jenkins job
Uses letsencrypt-aws to check the ELB certificate expiration date
and request a new certificate if the current one is valid for less
than 1.5 months.

Since letsencrypt doesn't support wildcard certificates at the moment,
the bash script wrapper around letsencrypt-aws automates the DNS host
discovery by looking for Route53 records pointing to the target ELB.

Once the certificate is downloaded it's renamed to match the ELB name.
This is needed to keep the IAM certificate ARN unchanged, since it's
used as a variable in CloudFormation updates, so if the ARN changes
without updating the digitalmarketplace-aws files next CloudFormation
run will attempt to restore old certificate. Keeping the name the same
should avoid this problem.

The preview job will run every day, but should be a no-op most of the
time since the certificate will be changed only if it's about to expire.