Add tasks to request a letsencrypt certificate for .beta domains
Requests a certificate from letsencrypt and sets up nginx config to
use it. Removes the steps that used to copy a certificate from the
credentials repo.

Uses letsencrypt standalone server on port 80, now that nginx isn't
listening to it anymore.

Letsencrypt certificates are valid for 3 months and there's no auto
renewal set up, so `--tags letsencrypt` needs to be rerun when they
expire. It should take care of requesting a new cert and restarting
nginx, but needs port 80 to be accessible for external IPs, which
can be done by creating `jenkins_letsencrypt_access` CloudFormation
stack if it doesn't exist.