Add notify-buyers-when-requirements-close Jenkins job
Runs the buyer notification script every day at 5am.
Jenkins is in UTC so setting the time few hours from midnight makes
it easier to reason about which brief users where notified.

Setting to run on preview and production for now (with a test mandrill
key for preview) so that we have an environment to check changes on.