Deprecate "SSP" from job names
We no longer have a "Service Submission Portal" at all, and haven't had one since G-Cloud 6. The acronym SSP is meaningless to many new members of the team (and some old ones). So it will be more useful to have a descriptive name for the tests on our build monitor as they fluctuate between red and green.