Download gdrive and create token file
Add an ansible task to get the `gdrive` command line tool and then
create the token file it needs to connect.

Note that ultimately the token file has more in it (see below),
but after testing locally, this is the smallest file that gdrive
can work with.

Example file once gdrive has been used:

```
{
  "access_token": "fake123stuff",
  "token_type": "Bearer",
  "refresh_token": "fake456stuff",
  "expiry": "2017-01-13T17:49:19.951483115Z"
}
```

Relies on the existence of a refresh token in the credentials repo.

Note as well that the gdrive binary is just being downloaded as a
file, so there's no upgrade command; we're just going to have to
update the download URL to get a new file.

[gdrive](https://github.com/prasmussen/gdrive)