Don't update/reinstall bundler if it's already installed
We're getting some occasional errors when installing bundler:

```
(Errno::ENOENT) No such file or directory @ rb_sysopen - lib/bundler/...
```

I suspect the problem is caused by a race condition between two jobs. I
think it happens when both functional tests AND smoke tests are about to
run. Before they were using different branches and different ruby versions.
Now they're using the same one, which means if they both try to run gem
install at the same time it'll modify the same files. My guess is at some
point gem install deletes old files before copying the new ones in and
that's when the other one can't find things in place.

Since the main thing we care about is that `bundler` has been installed
for the ruby version we're using we don't need to reinstall it each time
the job runs. Adding `--conservative` flag to `gem install` stops it
from reinstalling or upgrading the gem each time, as long as there's at
least one version already installed.

This means we're no longer running the latest bundler version available,
but it seems unlikely that we'd need to upgrade it without moving to a
newer ruby version.

This should still fix it for the majority of runs. But it also means this
could happen with other gems if we bump the version in the functional tests
Gemfile. Maybe we need to look into separating interpreters / gemsets if
this is still a problem later.