Tweak icon
A difficult choice this one because :chart_with_downwards_trend: is blue and :chart_with_upwards_trend: is red.

The red would indicate a problem, and our other job failure icons are mostly red, but then the downwards trend feels like a better representation of statistics failure so I'm going with that.