Add basic jenkins script to send production dos supplier emails
https://trello.com/c/qeY6Bl3U/253-set-up-the-jenkins-job-to-run-the-email-script-on-weekdays

Note, for the moment we are hardcoding in the list ID of the
team's test list just to be on the safe side whilst in QA that
no emails are unintentionally sent to real users.

The script will run every weekday at 8am.

For the moment we are only running on the production environment.
In time we may add further environments, especially when we start
trying to automate the user list updating.