Add choice parameters for lot slug, list id and number of days
For in case we wish to run the job manually.

Note, we still leave in using the test list ID by default until
we are happy with the job after the QA period. When we are happy
we can remove this line and it will revert to use production
users.