Replace individual app build jobs with a single 'build-app' job
Instead of watching application repos to trigger the build with a
separate job for each app we make the build job a separate common
step for all applications (similar to release-app).

The build job should be triggered by a separate pipeline for each
app that will tie together build, preview release and functional
tests.