Add a make variable to only update specific Jenkins jobs
Since there's no way to test jobs locally we're often applying them
to Jenkins during development, which makes it hard for more than one
person to work on them at the same time.

Replacing the wildcard in the task with a variable allows us to only
update the jobs matching the provided glob pattern.