Remove renamed or deleted job templates from the job builder
Our jobs tasks upload new version of templates on each run
to a directory on the Jenkins server, which is then used by
the jobs builder.

However, if a template was renamed or removed locally it will still
exist in the Jenkins directory and might cause issues with the job
builder (eg renaming a local job file will cause jobs builder to
complain about a duplicate job definition).

We could remove the remote directory before recreating it, but it
will mark all jobs as changed in the Ansible output.

Instead, we use rsync to delete any remote files that don't exist
in the local job_definitions folder without updating any of the
existing remote files.