Update build_application job to build specific releases
Build application job now requires a RELEASE_NAME to be set and the
corresponding tag/ref to exist in the application git repo. It's no
longer trying to parse the commit message to look for PR number.

As a result, it can build an artefact from any revision, not just
the current master branch.

Build application now checks if the release exists in s3 first and
quits early, without replacing it if it does.