Update PaaS release job to fit into the release pipeline
Removes support for automatically setting "RELEASE_NAME" from the
previous environment tag. "Promotion" should now be handled by the
full release pipeline which will call the release job with the
correct RELEASE_NAME.

Removing the tag parsing allows us to avoid cloning the git repo
altogether, which means the release job is only using the deployment
zip archive.

This also means we can no longer set the "deployed-to-*" tag here,
so we're moving this to the pipeline job for now.

This change also sets CF_HOME to the workspace directory, so that
multiple release jobs could run concurrently without the first one
logging out the others out of cf-cli.