Run functional tests after release to preview/staging
The new release pipeline for the PaaS didn't run functional tests, now
it does.

If functional tests fail, rather than fail the entire build you're now
presented with the option to retry, skip or abort. Retry will run the
functional tests on that stage again, skip will move on to the next
stage of the pipeline without rerunning the failing tests, and abort
will fail the entire build.