Wrap functional tests in milestones
By wrapping the calls to run the functional tests in milestones it
ensures that older builds that are waiting for input in the
`run_functional_tests()` function, will be aborted by jenkins if a newer
job completes the functional tests and hits the second milestone.

We could technically just use one milestone immediately before the
`run_functional_tests` call in most places, as the following stages
start with a milestone before user input. However, if extra stages were
added at a later date that didn't start with a milestone, this could
allow a build to proceed past the functional tests without aborting
earlier builds.