Pipeline jobs can't be disabled so removing the "disabled" key
This doesn't stop the job from being built OK, but gives a false sense of confidence having it in the job builder here.

Even with `disabled:true` here the job can be run on Jenkins at the click of a button.

Jenkins bug is here: https://issues.jenkins-ci.org/browse/JENKINS-27299