Prune images instead of deleting images with children
The script was returning with a non-zero status due to trying to delete
all images that didn't have the latest tag. All images includes the
images that make up the layers of the images we do want. You can't
delete these. If you try you get `unable to delete 0123456789abc (cannot
be forced) - image has dependent child images`

To fix this, we just delete those images that are not intermediate
images. This however leaves unused images which can be cleaned up with
prune.