Append jenkins user to jenkins group
Previously, when the jenkins user was added to the jenkins group it
wiped out any other group associations.

For jenkins to be able to use docker, it needs to be part of the docker
group.

If the jenkins `main` tasks are run, and not the `docker` tasks jenkins
will be removed from the docker group.

By appending, we can preserve the jenkins/docker association.