Decrypt ci ssh key before running tasks
The ci ssh key has been rotated and encrypted. Unfortunately we can't
use process substitution `<()` to decrypt the key, as ssh reads the file
more than once when establishing a connection.

This decrypts the key to a temporary file and the cleans it up after. An
issue is that if the jenkins job errors for any reason (for example
forgetting to activate your virtualenv), the key will be decrypted but
not cleaned up.