Use github.com instead of Enterprise
Repo's on github.digital.cabinet-office.gov.uk have been moved to
github.com. Jobs which use the old repos have been pointed at the new
reops.

This doesn't include the manual because whilst we can keep the manuals
repo private on github.com, if the job to build the pages is run we risk
leaking sensitive information. There is a story to check the manual for
this info.