Remove default tags value from `make jenkins`
Running the full playbook is not what we want most of the time and
it can get interrupted in an unexpected state (eg after deleting
Jenkins configs which it needs to do during first-time setup but
before copying the new ones).

Changing the behaviour to always require TAGS to be explicitly set
helps prevent accidental mistakes. It's still possible to run the
full playbook by specifying `TAGS=all`