Add job to upload a database dump to s3
The job will create an ssh tunnel to the database service on the PaaS,
and create a plaintext dump of the database.

The database is zipped and encrypted before being written to disc so
that it's not left unprotected at any point.

A script in the scripts repo is then used to upload it to an s3 bucket.