Add job for migrating and cleaning db dumps
The job will create a temporary postgres service and an app to ssh to it
through.

It will then import the latest production database dump to the new
service, clean it of sensitive data, take a dump of the new database and
then apply that dump to the TARGET_STAGE specified.

It will also upload it to a shared folder on google drive. If the
TARGET_STAGE selected is google-drive, it will not migrate the data to
another stage, just produce the dump in drive.

The default stage is staging, which allows us to set the job to run
every two weeks to automatically update staging with fresh data.