Index data and update alias
Once data has been migrated to a different stage it needs to be
migrated. There is already a job to do this which we can use. We set the
index name to the days date in keeping with how we currently do it.

Once that index has been created, the alias we use to denote which index
we're using needs switching to it. A script has been created in the
scripts repo to do that and delete old scripts with no aliases.

The next commit will be creating the `update_index_alias` job that's
being refered to here. That job will call that script.