Add check for alembic version and migrate if needed
It's possible that the target stage has more advanced migations that the
prod db dump. We can use a new script to check the dump version, and hit
the target stages status endpoint for it's version, and compare them. If
the dump version is less than the target version then we run the
migration job after applying the dump.