Don't alert smoke test failure if migrating data
When running the new data migration job the smoke tests for that stage
will fail whilst the new index is being created.

This changes the smoke tests jobs to a pipeline, and checks the build
status of the latest build of `clean-and-apply-db-dump`. If there is a
build currently being built it is used.

If the migration job is not running slack is notified of the smoke tests
failure. If the build IS running, but is not being built against the
same stage the smoke tests are running against then slack is also
notified.