Rename build-app job to build-image
Jenkins "build application" job is used to build a Docker image
from a tagged repository commit and upload it to Docker Hub.

Up until now it was only used to build application images, but the
job isn't really tied to a deployment process, so it can be used
to build any docker image. Renaming it to make this clear.