Add a job to run get-model-data CSV exports
Script is used to generate data analytics reports and save
them as CSVs.

It's using the scripts docker container to run the script and maps
a local folder as a volume to keep the reports in the workspace
for uploading to Google Drive.

Since Jenkins is used to build the scripts Docker image it should
always have the latest version available locally.

Job runs every Monday at 3am UTC.