Update all digitalmarketplace-aws jobs to Python 3
Activates python3 pyenv before installing requirements on all jobs
that are calling `make requirements` for digitalmarketplace-aws.

From #31:

digitalmarketplace-aws is now using python3 to set up virtualenv,
so we need to activate a Python 3 pyenv before installing requirements.

This requires hardcoding the python version in the job and any existing
venvs will keep using the linked interpreter versions until the job
workspace is manually cleaned.

For scripts we solved this by running them in a Docker container, but
it's not clear whether this would be easy to do for aws repo, so I'm
setting it up with pyenv for now.