Fix incorrect call to jenkins clean db job
We were calling jenkins to see if a migration was currently
being run. The original call being made assumed there was an
underlying job for both of our clean and apply db dump jobs.
However, there is no underlying job and we instead have two jobs
which both have their environment in the name e.g.
`clean-and-apply-db-dump-staging`.

Because the url we were calling did not exist because no job existed
under this name, the response was an HTML 404 page. This then meant
that jq could not parse this as it was not JSON. Our shell script
therefore exited and no slack notifications were ever triggered.

To fix this, we can simply call the clean and apply dump job for
the appropriate environment (and therefore don't need to work out
which stage the dump is happening on). Note, as production
doesn't have a clean and apply db dump job, we never have a case
where we should not notify slack. Further note, that we need
to define our `migrations_being_run` function in the jinja
if statement so the function is not defined in production. This
is because the function is evaluated when the script is run rather
than afterwards when we make a potential call to it.