Add dev's public ssh keys to authorized_keys
A temporary authorized_keys file is created, using the public key of the
shared ci ssh key as its base to ensure the shared key can always be
used to get access. The ci public key is stored in the creds repo next
to the private key.

The dev's public keys are then retrieved from the github api using the
id's stored in the jenkins vars file in the creds repo.

Once populated, the original authorized_keys file is replaced with temp
file.