Error if public key not found
If for some reason the public ssh key associated with the id listed in
the jenkins var file is not found, we return an error so we can fix it.

The -e flag to jq will exit with a non zero (4) status code if the final
output value is null or false. If .key can't be found then it will
error.

This requires pipefail to be set so the error isn't hidden by the awk
command.

Ansible will use dash be default which doesn't have pipefail, so we
need to set bash as the executable.