Conditionally create admin users for legacy func tests
The legacy functional tests require certain admin users to be set up.

The tests themselves don't do this.

If we wipe the preview db and restore from a cleaned version (as we do
with staging) we need to recreate those users. This is an annoying
manual chore.

Luckily we have a script to do this. Running it before the tests means
we don't have to remember to do this in the case of db restore.