Add jobs for rotating tokens
 ## Summary
Adds a new pipeline that allows for easily rotating data/search API
tokens in a given environment. Also updates the existing job for
releasing an application to paas to allow building without specifying
a release version. In this case, it will determine the existing release
by finding the app's primary route (via CF CLI) and hitting its
`_status` endpoint to find the current release tag. After doing that, it
will re-release that same version to PaaS, allowing new configuration to
be pulled in. The detection and deployment happens within the specific
app and stage's Jenkins lock, meaning it is effectively atomic and so
asynchronicity should not be a concern.

 ## Ticket
https://trello.com/c/4crGe6F0/169