Add jenkins job for notifying suppliers
 ## Summary
When applications for a framework close, we need to email all supplier
users to indicate whether or not they made a successful application for
the framework. We have a script that sends these emails that has
historically been run on a developer's machine. Since this script can
run for a long time, ideally it should run somewhere else, so this adds
a Jenkins job that will run the script.

 ## Ticket
https://trello.com/c/pgCNeI7P/126