Remove Jenkins aws config file
We had three scripts which were using an aws config file for
credentials. The credentials in the file belonged to a user called
`deploy` which was a hangover from when we used ElasticBeanstalk. That
user has quite a lot of permissions and we should probably get rid of
them.

The Jenkins box has its own credentials by way of an instance profile.
If we need Jenkins to interact with AWS we should probably just give
this instance profile (the role 'jenkins') the correct permissions,
rather than maintaining two different credentials.

There's a PR up on the AWS repo to adjust the jenkins role policy to
allow these extra permissions for reading/writing to certain S3 buckets.