Use `become` instead of `sudo`
Using `sudo` was deprecated. New Anible privelage escalation uses
`become`. By defining it in the playbook like we do, it means that all
tasks will be run as root. This means that we can remove it's use in the
tasks, except where we need to become a user other than root.

This also remove `gather_fact` as this is the default.