Refactor plugins tasks to use correct credentials
As it was, these tasks would run successfully the first time the box was
set up, but then fail on subsequent runs. This is because when
completing an initial setup, the authorization strategy we're using is
changed _after_ the plugins are initially installed.

This changes the way that we can authenticate against the Jenkins API.
As we move to using the Github OAuth app, on subsequent runs (if we were
upgrading our plugins for example) we need to authenticate with a Github
User. We have one!