Remove letsencrypt
Rather than using letsencrypt, we can get a free certificate from aws.
There's a PR on the terraform code to do this, by moving the new Jenkins
instance to behind and ELB. The main motivation for doing this was the
letsencrypt limit of 20 certificates a week. This sounds like a lot, but
when developing a new Jenkins/Ansible setup, it can really get in the
way. Plus it means we don't have to remember to renew every 90 days.

The ELB terminate our tls and forwards the requests on to nginx on the
jenkins box via http.