Remove task to delete config
This only runs when running ansible without any tags. I'm removing it
because it causes errors if you run ansible again withou any tags. On a
second run through, Jenkins is already installed and so the config isn't
replaced. This causes errors with csrf when getting a crumb for plugin
installation.

Also, this commit moves the ordering of the tasks around Jenkins is
setup all in one go.