Add awscli and pyenv to tools
AWS cli is neede for the database backup job.

pyenv was removed a while ago from the ansible but was was never removed
from the system. We have a couple of jobs that use it. Maybe we should
look into updating those jobs to not use it. But for now, we need it to
run them.