Use correct GPG2 config
We need GPG to allow loopback pin entry so that it can be used in a non
interactive manner. The version that we have (2.1.11) doesn't have it
enabled by default. The next version (2.1.12) does, but apt gives us
2.1.11.

We already had the correct config file for this, but for some reason we
weren't copying it to Jenkins.