Prevent ssh connection from timing out after 60 seconds
Default timeout for an idle connection was 60 seconds. If a task took
longer than 60 seconds to complete (looking at you, install pyenv) the
ssh connection was terminated by the server which stopped the playbook
from executing.

By adding these lines to the sshd config we can prevent this.

`ClientAliveInterval 45` makes the server send a null packet every 45
seconds which keeps the connection going

`ClientAliveCountMax 80` is the maximum number of times this null packet
will be sent. 80 gives one hour before the connection is dropped.