Add `copy_old_builds.sh` to playbook
If we're creating a new Jenkins box we're probably going to want to copy
the old builds from the outgoing box. There's a dead simple script to do
this using rsync. We can add it to the Jenkins build so it's available
without having to add it to the box later.

To use it, the ssh key being used will need filling in, as well the
hostname of the old server.

There didn't seem to be a great place to add this script copy, so I
created a new task. It needs to be run after the config task as this is
where the jenkins data directory is created with permissions.