Give briefs indexing job default frameworks
Previously the job was run with no frameworks listed. This had the
effect of indexing all DOS frameworks (the api returns all frameworks if
a specific framework is not specified).

This is fine if just indexing, but causes an issue if creating an index.
This is due to a new check which ensures the mapping specified for the
new index is correct for the frameworks being indexed into it. This
check fails and stops the indexing script if you don't supply the
frameworks.

This change defines a default value for DOS frameworks for indexing, in
the same way as we currently do for G-Cloud indexing. This means the job
can still be called without providing a value for frameworks and
behaviour will be maintained. It also means that the new check will
pass.