Create a new pipeline job to release all the apps
When inserting new credentials into the credentials repo we have to manually release all the apps, being _very_ careful not to miss anything, or use the rotate tokens job and manually release the route. Neither of these situations is very desirable and they are prone to human error. This job releases all the jobs on a given stage ensuring they all have a fresh version of the credentials.