Fix bug in test_templates.sh
`test_templates.sh` was not using the global defaults because it was
only parsing one jobfile at a time.

This commit changes the JJB run to include all templates at once, which
more closely reflects the behaviour of the Jenkins Ansible role.

It does mean the output is slightly less helpful though.