Add sshd rule to Jenkins box
Use the willshersystems.sshd role to configure sshd on the Jenkins box.
This commit migrates the config changes previously made in the
bootstrap.yml playbook in the Jenkins role, and also includes the
"PasswordAuthentication no" tweak introduced by the AWS machine image.

This commit results in ansible reporting changes as the sshd_config file
generated by willshersystems.sshd is not line for line exactly the same
as the Ubuntu default; however, it is functionally the same (i.e. the
configuration options are the same).