Send DOS opportunities email job is now framework agnostic
Changes to the script mean that all newly published briefs are
included and emailed to the relevant Mailchimp list of suppliers,
regardless of framework. This means that on the crossover day
where buyers may have published both DOS3 and DOS4 briefs, the
script can create separate campaigns (per lot) for both frameworks.