Add visual regression tests job to staging
We want to be able to check whether changes we are releasing will cause
visual regressions.

This commit adds a visual-regressions-staging job, and sets it to be
automatically run as part of the release pipeline for an app.

We also make sure visual regressions are run against staging when
restoring a database dump.