Run visual regression tests after applying database dump
Restoring a database dump to a stage can change the content for account
pages of test users used in visual regression tests. This will cause the
visual regression tests to fail next time they are run.

Rather than creating false positives against a merge, we should run and
approve the visual regression tests against the stage immediately after
the restore has been done. This commit adds a stage to the pipeline that
runs the visual regression tests and then automatically approves the
changes.