Add upper limit on number of builds to keep for tests
We've had issues where we've run out of space on the Jenkins data
volume due to a large number of test runs being retained. This seems to
be because a large number of test runs took place in a short amount of
time. We have a limit on the number of days we keep logs and artifacts
for test runs, but not on the total number kept.

This commit adds a limit on the total number of builds retained for
functional test and visual regression test jobs; this should help reduce
the likelihood of running out of disk space in future.