Add random password for PostgreSQL container
Our clean_and_apply_db_dump job on Jenkins has stopped working because the
postgres Docker image expects an environment variable POSTGRES_PASSWORD.

This commit adds a command to the prepare step to generate a random
password and adds it to the environment. We also set PGPASSWORD in the
environment for the psql command.

Depends on alphagov/digitalmarketplace-aws#628