Fix notify-suppliers-of-framework-application-events
Jinja was unable to parse this job template because we wanted to access
a dictionary variable with a run-time Jenkins parameter.

For now I've hacked it to use awk and some not very nice looking choices
for the parameter; in future we could look at using a Jenkins DSL script
instead.