Jenkins job for publish-draft-services script
This job should be able to switch roles using the new AWS profile configuration.

The draft-ids parameter has been left out of the script for now, as it refers to a file upload. That parameter is likely to be used on a failure re-run and should probably be handled manually rather than on Jenkins anyway.