Update mark-definite-framework-results to log Jenkins user in audit logs
We want the Jenkins user who starts the mark-definite-framework-results
to be recorded in our audit log, so we can see later who made a change.

By default the mark-definite-framework-results script uses the unix
username, which on Jenkins is not very useful. Instead we use the build
tag and build user envvars to make a more useful message.