Move config for Jenkins views to Jenkins Job Builder
Use Jenkins Job Builder to manage Jenkins views, instead of generating
the config with Jinja.

This has the advantage that running `make jenkins TAGS=jobs` will now
update lists as well as jobs.

Jenkins Job Builder currently does not support the Build Monitor view
plugin, so for now we leave that as is.