Make common tags into their own targets
We recently had an incident when a developer accidentally ran `make jenkins` using the wrong tag and caused Jenkins to restart at an inopportune moment.

Add targets for the most common tags. This should make it a bit more difficult to do the wrong thing in future.

Leave the existing target so that people can still choose to run uncommon tags. This avoids us needing to maintain a vast list of rarely-used targets.

We'll need to update the team manual after this gets merged.