Move extra variable calculation into bash
Previously, we constructed `EXTRA_VARS` in make, which bash interpreted (and passed to ansible) as a single string. Ansible did not recognise this as a valid extra var.

Now, we construct `EXTRA_VARS` in bash, where we can use a bash array to ensure that it all works correctly.

Also replace short options with long options for legibility.