Update Docker directory config
`graph` was deprecated in Docker v17.05[[1]] and is replaced by
`data-root`[[2]].

We're currently on v20.10.5 so Docker's fallen back to
using the default location of `/var/lib/docker` on the smaller drive.
This is likely the cause of the P4 incident where Docker used up all
the available disk space and failed to build new images.

[1]: https://docs.docker.com/engine/deprecated/#-g-and---graph-flags-on-dockerd
[2]: https://stackoverflow.com/questions/24309526/how-to-change-the-docker-image-installation-directory/50217666#50217666