add reference and optional override
In order to use the reference/retry functionality of the email engine we need to persist the log file. This mounts the logfile generated inside the docker image /tmp folder to the Jenkins workspace directory.

It also adds a default reference of the Jenkins Build Number with an optional override to allow for resuming previously failed builds