Notify Slack if the code backup job fails
Previously we didn't get a notification in Slack if the code backup jobs
failed. This could cause problems if we needed to use them and found
out at that point that there was a problem.

Uses a [publisher](https://docs.openstack.org/infra/jenkins-job-builder/publishers.html)
to trigger an action after the build completes. See the
[`export_supplier_data_to_s3`](https://github.com/alphagov/digitalmarketplace-jenkins/blob/main/job_definitions/export_supplier_data_to_s3.yml)
job for an example of how we already use this syntax.