Throttle app pipelines to 4 concurrent jobs
We just had Jenkins deadlock. This is because we had a large number of app pipeline jobs all in progress. However, they were all blocked on a deploy job, so could not make progress. However, all the deploy jobs were blocked on other jobs, mainly notifying Slack. However, the jobs in this third tier couldn't be scheduled because all the executors were full of blocked jobs in the first and second tier. We had to cancel a bunch of jobs to clear the deadlock.

Introduce throttling on the app pipelines. Now, no more than 4 can run at once. This means that in the worst case, only 12 executor slots will be occupied by jobs related to app pipelines. So other jobs will still be able to run. I think this should make it much harder to accidentally deadlock Jenkins.

We previously tried doing this in 5c8736f. That didn't work because the throttling was done on too wide a scope, so picked up pipelines that weren't even trying to make progress. I think this approach should work, by only throttling just before we want to allocate an executor node.

Let's see if this works, and quickly revert if not.