Run functional tests using headless Firefox
Our tests are currently rather unreliable. I have a hypothesis that this might be PhantomJS-related and that switching to Selenium might improve things.

Do an experiment to find out whether this is the case. Build on https://github.com/alphagov/digitalmarketplace-functional-tests/pull/901 to switch the functional test jobs over to use Selenium with headless firefox. Install firefox so that this works.

My plan is to deploy it and see what happens. Should be quick and easy to revert if it breaks things.