Send all Slack notifications to #dm-2ndline
This is a temporary measure. Our CCS techops friends only have access to #dm-2ndline. So send all Slack notifications there so that they can see them. This will make that channel a lot more spammy, but I think that's an acceptable price to pay.

We should be able to revert this change once we've switched over to a CCS Slack (or equivalent) where we can use more than one channel.